Move kernel pieces of cumsum/cumprod into unfied file.
1. Fixing compilation error in BUILD_SEPARATE_OPS mode.
2. Although we aligned with file naming of PyTorch in-tree, putting host only code in a file under ../xpu/sycl doesn't align with our design.
